groot explicit book: chain rung, attention gate, and the projection race
The explicit seat book now walks the same ladder the receipts drive:
the DiT stack binds the fused NVFP4 chain first and the per-seat
declarations yield their claim only when the chain answers; the
backbone attention interface rides the recorded band decision; and
the linear projection family races the host on the calibrated shape
before it seats - on one device the separate bias and quantize
launches lose to the host's bias-fused GEMM, on another they win,
and the same declaration gives each box its answer.

The runner learns that a stock baseline is a receipt, not a rerun:
FRT_SKIP_STOCK skips its compile and FRT_STOCK_MS carries the
archived number. The pi05 vision-key shim casts to the parameter
dtype it retargets.

+++ examples/pi05.py
@@ -140,7 +140,7 @@ def _shim_vision_state_keys(policy) -> int:
         flat = k.replace(".vision_model.", ".")
         for mk in want:
             if mk.endswith(flat) or flat.endswith(mk):
-                fixed[mk] = v
+                fixed[mk] = v.to(msd[mk].dtype)
                 break
     if fixed:
         policy.load_state_dict(fixed, strict=False)
